add: payment & promocode services communication with ride service

+++ promocode-service/src/main/java/com/dulfinne/taxi/promocodeservice/dto/response/PromocodeUsageResponse.java
@@ -10,6 +10,6 @@ public record PromocodeUsageResponse(
     String code,
     String username,
     Instant usageDate,
-    Integer rideId
+    Long rideId
 ) {
 }
